requirements.txt
adafruit-blinka
adafruit-circuitpython-neopixel
opencv-python
paho-mqtt
flask
pyyaml